Written Analysis
# 1-Overview of Project: Explain the purpose of this analysis.
The purpose of this analysis is to assist Louise to achieve funding goals on Kickstarter using statistics and data analytics on past projects.  The goals is to provide her with bar charts, line graphs and other analytics in excel so that she can determine when and which projects are most likely to succeed in reaching funding goals.

2-Analysis and Challenges: Explain how you performed your analysis using images and links to code, as well as any challenges you encountered and how you overcame them. If you had no challenges, describe any possible challenges or difficulties that could be encountered.

I performed my analysis using a number of tools learned in the module.  I used statistics, specifically averages, outliers to determine how likely her project is mostly to be funded.  I used bar charts to suggest which months she should consider launching her funding projects.  We compared several projects to confirm and project if funding goals will be reached or failed.  I did not have any challenges using the tools learned.  Challenges that could be encountered will be trying to address data where we were finding ERRORs in our calculations.

3-Answer the following questions in complete and coherent sentences.
•	What are two conclusions you can draw about the Theater Outcomes by Launch Date?
o	The first conclusion I drew about Theater Projects by launch date is that they are generally more successful on May and June
o	The second conclusion is that December and January are not ideal months for to launch a theater funding campaign.  They have an equal chance of failing 
o	What can you conclude about the Outcomes based on Goals?
Based on the charts we generated, projects with a goal of raising over $50,000 has a 58% chance of failing with only a 19% chance of being successfully funded.
•	What are some limitations of this dataset?
o	The dataset does not give a lot of information on the motivation of the donors and their incentive to fund a project.  Therefore, it is hard to assess why donors support a project since the data on donor satisfaction is not available. 
o	Looking at the number of projects launched on a monthly basis, about half of them failed every month.  We do not have any insights about the causes of their failures.
•	What are some other possible tables and/or graphs that we could create?
o	There some many other tables and/or graphs that could be created to get more insights on the data.  For instance, I would create a pie chart to visualize which project category got the biggest slice or was the most successful among all of the projects on the Kickstarter platform. 
o	I could also use a scatter plot to visually illustrate the relationships between two variables.  For instance, I could use the scatter plot to see subcategory projects that were canceled on the Kickstarted platform.